backend/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
numpy
python-dotenv
pydantic
